// src/data/questions.generated.ts
export const interviewQuestions: InterviewQuestion[] = [
  {
    "id": 1,
    "category": "重要十大問題(1~10至少問一題)",
    "question": "上一份工作離職的原因/為什麼會想轉換跑道？",
    "difficulty": "核心必練",
    "tags": [
      "experienced",
      "motivation",
      "top10"
    ]
  },
// ...
];